refactor(switch): move from experimental to components
BREAKING CHANGE: Switch has moved to src/components/switch
