+++ src/components/form/form.stories.mdx
@@ -12,7 +12,7 @@ import DateInput from "../../__experimental__/components/date";
 import { Select, Option } from "../select";
 import { Checkbox } from "../checkbox";
 import Hr from "../../components/hr";
-import Switch from "../../__experimental__/components/switch";
+import Switch from "../switch";
 import Button from "../button";
 import DialogFullScreen from "../dialog-full-screen";
 import Dialog from "../dialog";
