+++ src/components/switch/__internal__/switch-slider.style.js
@@ -1,8 +1,9 @@
 import styled, { css } from "styled-components";
 import PropTypes from "prop-types";
+
 import baseTheme from "../../../style/themes/base";
 import SwitchSliderPanel from "./switch-slider-panel.style";
-import StyledValidationIcon from "../../../components/validations/validation-icon.style";
+import StyledValidationIcon from "../../validations/validation-icon.style";
 
 const StyledSwitchSlider = styled.span`
   ${({ checked, disabled, size, theme, error, warning, info }) => css`
