+++ src/components/switch/switch.style.js
@@ -1,14 +1,14 @@
 import styled, { css } from "styled-components";
 import PropTypes from "prop-types";
 import { margin } from "styled-system";
-import baseTheme from "../../../style/themes/base";
-import FieldHelpStyle from "../../../__internal__/field-help/field-help.style";
-import HiddenCheckableInputStyle from "../../../__internal__/checkable-input/hidden-checkable-input.style";
-import { StyledLabelContainer } from "../label/label.style";
-import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
-import StyledSwitchSlider from "./switch-slider.style";
-import StyledValidationIcon from "../../../components/validations/validation-icon.style";
-import { FieldLineStyle } from "../form-field/form-field.style";
+import baseTheme from "../../style/themes/base";
+import FieldHelpStyle from "../../__internal__/field-help/field-help.style";
+import HiddenCheckableInputStyle from "../../__internal__/checkable-input/hidden-checkable-input.style";
+import { StyledLabelContainer } from "../../__experimental__/components/label/label.style";
+import { StyledCheckableInput } from "../../__internal__/checkable-input/checkable-input.style";
+import StyledSwitchSlider from "./__internal__/switch-slider.style";
+import StyledValidationIcon from "../validations/validation-icon.style";
+import { FieldLineStyle } from "../../__experimental__/components/form-field/form-field.style";
 
 const StyledSwitch = styled.div`
   ${({
